Crate penumbra_sdk_proof_params

Modules

convert

delegator_vote

nullifier_derivation

output

spend

swap

swapclaim

Structs

LazyProvingKey

A wrapper around a proving key that can be lazily loaded.

Constants

GROTH16_PROOF_LENGTH_BYTES

The length of our Groth16 proofs in bytes.

Statics

CONVERT_PROOF_PROVING_KEY

Proving key for the convert proof.

CONVERT_PROOF_VERIFICATION_KEY

Verification key for the convert proof.

DELEGATOR_VOTE_PROOF_PROVING_KEY

Proving key for the delegator vote proof.

DELEGATOR_VOTE_PROOF_VERIFICATION_KEY

Verification key for the delegator vote proof.

NULLIFIER_DERIVATION_PROOF_PROVING_KEY

Proving key for the nullifier derivation proof.

NULLIFIER_DERIVATION_PROOF_VERIFICATION_KEY

Verification key for the delegator vote proof.

OUTPUT_PROOF_PROVING_KEY

Proving key for the output proof.

OUTPUT_PROOF_VERIFICATION_KEY

Verification key for the output proof.

SPEND_PROOF_PROVING_KEY

Proving key for the spend proof.

SPEND_PROOF_VERIFICATION_KEY

Verification key for the spend proof.

SWAPCLAIM_PROOF_PROVING_KEY

Proving key for the swap claim proof.

SWAPCLAIM_PROOF_VERIFICATION_KEY

Verification key for the swap claim proof.

SWAP_PROOF_PROVING_KEY

Proving key for the swap proof.

SWAP_PROOF_VERIFICATION_KEY

Verification key for the swap proof.

Traits

DummyWitness

This trait characterizes circuits which can generate constraints.

ProvingKeyExt

VerifyingKeyExt

Functions

generate_constraint_matrices

Generate constraint matrices from a circuit type.

generate_prepared_test_parameters

A variant of generate_test_parameters which spits out a verifying key with some precomputation.

generate_test_parameters

Generate parameters for proving and verifying, for tests.