penumbra_sdk_proof_params
Trait DummyWitness

pub trait DummyWitness: ConstraintSynthesizer<<Bls12_377 as Pairing>::ScalarField> {
    // Required method
    fn with_dummy_witness() -> Self;
}

This trait characterizes circuits which can generate constraints.

Required Methods

fn with_dummy_witness() -> Self

This will create a circuit with dummy witness values, for constraint synthesis

(The reason this is needed is because constraint synthesis encapsulates both the act of generating the constraints, but also that of providing the witness values when proving). ((For the record, I am not a fan of this)).

Object Safety

This trait is not object safe.